% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_up_bank_auth_token.R
\name{get_up_bank_auth_token}
\alias{get_up_bank_auth_token}
\alias{get_up_bank_auth_token_from_dotenv}
\title{Get UP Bank auth token}
\usage{
get_up_bank_auth_token(call = rlang::caller_env())

get_up_bank_auth_token_from_dotenv(call = rlang::caller_env())
}
\arguments{
\item{call}{}
}
\value{
UP Bank Auth Token
}
\description{
Will retrieve the auth token. It will get it from the following places:
\enumerate{
\item Sys environment
\item .env file
}
}
\details{
It will throw an error if no
}
