% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_decoders.R
\name{json_decoder}
\alias{json_decoder}
\title{Decode Json}
\usage{
json_decoder(res)
}
\arguments{
\item{res}{response from the up bank API}
}
\value{
tibble containing the contents of the response json.
}
\description{
These functions decode the response json from a range of Up Bank API end points. It uses S3 to determine which decoder to use.
}
