% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_up_accounts.R
\name{read_up_accounts}
\alias{read_up_accounts}
\title{Read Up Accounts}
\usage{
read_up_accounts(
  account_id = NULL,
  account_type = c("SAVER", "TRANSACTIONAL", "HOME_LOAN"),
  ownership_type = c("INDIVIDUAL", "JOINT")
)
}
\arguments{
\item{account_id}{Optional; account_id to request from API. If its not provided the function will return all accounts for the authenticated user.}

\item{account_type}{Optional; account_type to filter request. It must be one of 'SAVER', 'TRANSACTIONAL' or 'HOME_LOAN'}

\item{ownership_type}{Optional; ownership_type to filter request. It most be one of 'INDIVIDUAL' or 'JOINT'}
}
\value{
\code{tibble::tibble()} containing the requested accounts data
}
\description{
This function is used to extract accounts data from the Up Bank API. It has three optional parameters that can be used to filter the query that is sent to the Up Bank API.
}
