% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_up_transactions.R
\name{read_up_transactions}
\alias{read_up_transactions}
\title{Read Up Transactions}
\usage{
read_up_transactions(
  transaction_id = NULL,
  status = c("HELD", "SETTLED"),
  since = NULL,
  until = NULL,
  category = NULL,
  tag = NULL,
  formatter = "simple"
)
}
\arguments{
\item{tag}{}
}
\description{
Read Up Transactions
}
